requirements.txt
crewai
crewai_tools
python-dotenv  # Load environment variables (API keys)